chore(plugin-k8s): clean up plugin-k8s directory structure
Fixes #515
