+++ plugins/plugin-k8s/src/lib/controller/abbreviations.ts
