+++ plugins/plugin-k8s/src/lib/model/states.ts
@@ -18,7 +18,7 @@ const debug = require('debug')('k8s/states')
 
 import repl = require('@kui-shell/core/core/repl')
 
-import { maybeAsDate, TryLaterError } from './util'
+import { maybeAsDate, TryLaterError } from '../util/util'
 
 /**
  * Resource status/states
