+++ plugins/plugin-k8s/src/lib/view/formatEntity.ts
@@ -18,8 +18,8 @@ const debug = require('debug')('k8s/util/formatEntity')
 
 import eventBus from '@kui-shell/core/core/events'
 
-import { flatten, isDirectory, toOpenWhiskFQN } from '../cmds/util'
-import { State, States, FinalState, endWatchWithConflict, watchStatus, rendering as stateRendering } from '../cmds/states'
+import { flatten, isDirectory, toOpenWhiskFQN } from '../util/util'
+import { State, States, FinalState, endWatchWithConflict, watchStatus, rendering as stateRendering } from '../model/states'
 
 /**
  * Make a kube context attribute
