+++ plugins/plugin-k8s/src/lib/view/formatTable.ts
