+++ plugins/plugin-k8s/src/lib/view/helm-status.ts
