+++ plugins/plugin-k8s/src/lib/view/redact.ts
