+++ plugins/plugin-k8s/src/lib/view/registry.ts
